refactor: co-locate ScrollOnNav with the other components
It is a component, not a utility, and its sibling ScrollToTop already lives
in components/. src/utils/ is now empty and gone.
